[PORT FROM R44C] libcamera2: remove video height aligning
BZ: 207062

This patch removes the video height aligning, which has not been
working for 1080p anymore. The code which made the height from
1080 to 1088 (16 aligned), has been countered with the height
limiter code just under it, producing a warning print:
"invalid recording height 1088, override to 1080"

Thus the aligning code has not been doing anything useful for a
long time, and this patch removes it to get rid of the unnecessary
warning.

Change-Id: I54ba70656f67cf1f0202dc806e4b30d6aa566c77
Signed-off-by: Ukri Niemimuukko <ukri.niemimuukko@intel.com>