[PORT FROM R44C] libcamera2: REVERTME: skip the first frame for continuous shooting
BZ: 204423

When set number_capture to -1 for offline capture we get 3-4 old frames
in the second time continuous shooting. In FW, there are 3 places hiding
buffers from previous session:
1. Tagger buffers
2. Buffer between tagger and capture pipe
3. Capture pipe

First 2 are easy for FW to resolve but for item 3, it's very hard to
get this buffer back when it has entered the capture pipe.

HAL adds this workaround to skip one old frame in the second time
continuous shooting.

REVERTME when FW bug 3039 is fixed.

Change-Id: Ifaade7e30352bac9cae4a5a60f48c6d12d44468b
Signed-off-by: Xie,Hongbiao <hongbiao.xie@intel.com>