[PORT FROM R44C] libcamera2: start ext-ISP CAF right after mode setting
BZ: 204796

For contiuous auto-focus the AF needs to be started right after mode
setting, as it needs to be run continuously. In contrast to AF done with
autoFocus() Android API call, which is user controllable. This patch
enables AF when other AF modes than the ones requiring autoFocus()
call are set.

Change-Id: I2d9e89c31efad718a631c40965e3f46d43dadeb1
Signed-off-by: Lassi Väätämöinen <lassix.vtminen@intel.com>