[PORT FROM R44C] libcamera2: REVERTME: SW encode images with non-8-aligned height
BZ: 209333

Our HW JPEG encoder does not properly handle images whose
height is not aligned to 8, for example, our wide 6MP image
(3264x1836). HW encoder leaves empty data at the bottom of
the image that causes a green line after decoding for a thumbnail.

Also, when padding the YUV data to be 8-aligned before feeding to the
HW encoder causes malformed image data to be produced at the bottom, so
adding extra padding does not seem usable at the moment either.

This workaround introduces using SW encoder for images with height
not aligned to 8.

Change-Id: I2b149e436b8921a1b2a0dd0e217f6c0040042b42
Signed-off-by: Lassi Väätämöinen <lassi.i.vaatamoinen@intel.com>