[PORT FROM R44C] libcamera2: flush buffers before HW jpeg encoding
BZ: 207735

This patch starts to pass the cache flush status of the buffer
that is to be encoded, to the encoding code in PictureThread.
Then PictureThread can choose between either flushing the buffer
or using the sw encoder for those buffers, that are not flushed.
Current solution is to flush the buffer and hw encode, except
panorama pictures which often have dimensions that the hw encoder
does not support.

Generic flush code is added to MemoryUtils and the cache line
size is added to PlatformData.

Change-Id: I6d5a0d69c4ad93d48a684ddb9f43a77e8d021fd1
Signed-off-by: Ukri Niemimuukko <ukri.niemimuukko@intel.com>
Signed-off-by: Lassi Väätämöinen <lassi.i.vaatamoinen@intel.com>