[PORT FROM R44C] libcamera2: WORKAROUND: alloc 128-aligned sizes for RAW capture
BZ: 207736

For RAW capture, we need to allocate buffers that have BPL aligned
to 128 and also the frame size towards ISP needs to be configured
to sensor output frame size, instead of application-set size.

With this patch we select the closest match for application-set picture
size from the sensor output sizes and allocate buffers matching to this size.
Also, BPL will be aligned to 128, if needed.

NOTE: the BPL adjustment can be removed, once CSS FW is
fixed to report correct BPLs.

Change-Id: Ib3dc2f59673e2ff5e80fcf0a87c5e9326920b92e
Signed-off-by: Lassi Väätämöinen <lassi.i.vaatamoinen@intel.com>