[PORT FROM R44C] libcamera2: stop still AF when entering pre-flash sequence
BZ: 209333

The still AF sequence must be stopped when entering pre-flash
sequence. This is needed, because for pre-flash sequence we
enable the AF lock, which stops the AF handling in 3A.

Another reason is the fact that the flash sequence is run on
'frame'-event, but still AF sequence runs on 'stats'-event.
In flash sequence the 'stats'-event is ignored, which interrupts
the AF sequence handling, under right circumstances. Namely, in
case the pre-flash sequence is started before AF finishes. E.g.,
when takePicture() is called right after autoFocus() API call.

This patch addresses the issue by stopping still AF sequence when
flash sequence is entered, to avoid invalid operation of AIQ,
which caused the pre-flash sequence to go on indefinitely.

Change-Id: Ibd77d56ec57ff51b0b60652aefa4a21044851994
Signed-off-by: Lassi Väätämöinen <lassi.i.vaatamoinen@intel.com>