libcamera2: change face callback memory allocation
BZ: 207677

The memory handling was static previously, and face data transfer between
PostProcThread - CallbacksThread was handled by memcopying data. This
required introducing a mutex in CallbacksThread, which causes deadlock
issues for example when Android API call for setParameters() has been issued
and face detection is running.

As setParameters() in the framework takes the service-level lock, us
using the lock in Callbacks class when sending face callback to application
caused a deadlock, if the incoming setParameters() call triggered a preview
restart. Preview restart includes stopping face detection, and as FD stop is
done synchronously, the setParameters() call got blocked indefinitely as a
consequence.

To fix this issue properly, this patch modifies the face callback data
allocation and passing so that each message will have dedicated face
metadata, which can be passed to CallbacksThread. CallbacksThread
will then generally deallocate the metadata, once done with the callback.
This helps us to get rid of the extra lock, which caused the deadlock, as
now the member variable handling is moved to CallbacksThread context.

Also added flushing mechanism to flush face metadata messages in
CallbacksThread from other threads.

Change-Id: I4be8e67575f04e0c5b8c5bf65e5f4f87c8b873cb
Signed-off-by: Lassi Väätämöinen <lassi.i.vaatamoinen@intel.com>