libcamera2: reset status when failed to get isp device path
BZ: 211533

In Bug 180111, we have an algorithm in CameraHAL which
identifies which subdevice we are using, so that we don't
have to hardcode such value in camera_profiles.xml anymore.

In some cases, when we failed to get isp device path, we
should fall back to tranditional method XML config.  However,
status from mSensorHW->getIspDevicePath was not updated, and
therefore even if we used the value from camera_profiles.xml,
we would still return Error initializing ISP.

Solution:
When we failed to get isp device path and fall back to
tranditional method XML config, we should reset status = NO_ERROR.

Change-Id: I7d3747e110d98b2ddc3b3bc3c653a353ad602f74
Signed-off-by: Lai, Jim <jim.lai@intel.com>