libcamera2: allocate snapshot buffers from graphics
BZ: 199580

This patch reverts back to allocating snapshot buffers from
graphics, because in devices with 4GB of memory the heap
allocation address may be from above 2GB region, which does not
work for the hw encoder. Graphics allocations are from below
2GB addresses.

The change is done only for devices with more than 2GB of memory,
in order to avoid issues seen previously with the devices that
have very little memory.

Change-Id: Ifef438f284410cf58a043ce053cbc3ef411bce4c
Signed-off-by: Ukri Niemimuukko <ukri.niemimuukko@intel.com>