libcamera2: detect repeating isp timeouts
BZ: 213960

Sometimes isp timeouts seem to keep happening despite of the
preview recovery mechanism. This patch adds a timestamp to
the timeout, and HAL will then detect if the timeout is
repeating. If it is repeating, the camera is stopped,
deinitialized and the mediaserver is restarted, to avoid
the need to reboot.

Change-Id: Idab25cee92c5e3b9eb39d0ee38109f1a66d9c6bf
Signed-off-by: Ukri Niemimuukko <ukri.niemimuukko@intel.com>