libcamera2: Stop and delete SensorThread at exit
BZ: 215454

SensorThread was never deleted and SensorLooperThread was not
stopped when the listener count dropped to zero. This patch fixes
these. This seems to fix the known cases of SensorLooperThread
starting to run in a busy-loop kind of fashion.

Additionally, this patch adds a last resort timestamp mechanism to
the SensorLooperThread, so that if it still starts to busyloop for
some reason or the other, a warning will be printed and the thread
will be throttled down by forced sleeping.

Change-Id: I376c2543c74dc9d99412415e5d7b1f8a6641115c
Signed-off-by: Ukri Niemimuukko <ukri.niemimuukko@intel.com>