libcamera2: Remove hard coded JPEG capture bpl
BZ: 215674

In JPEG capture, BPL depends on MIPI packet size used by ext ISP,
which is different for different external ISP variants. The driver
should be able to return the correct value always and HAL does not
need to be aware of it.

This patch also cleans up old CTP hacks for display BPL. These are
not needed in current platforms.

Change-Id: I0b9f6270d1be94a1a078314374f72718bd5c7d6a
Signed-off-by: Jarno Jakonen <jarno.jakonen@intel.com>