libcamera2: fix for the face callback deadlock
BZ: 214144

Locking in the CallbacksThread context caused a situation where
it is possible for the lock to stay held while a callback is being
blocked from entering Camera Service because some API call is trying
to simultaneously execute inside Camera HAL.

In practice the sequence was like:

-PostprocThread doing face callback, this tries to take the HAL lock
-CallbacksThread sending previous face callback, holding the HAL lock
 and trying to get CameraService lock
-API call towards HAL trying to turn smile functionality off, holding
 Camera Service lock, trying to do synchronous message towards
 PostProcThread, which is blocked by trying to get HAL lock

-> deadlock.

Solution is to not take locks inside CallbacksThread context. In order
to achieve this, the face data needs to be copied with messages. This
grows the CallbacksThread message size to 1.5kB, but since messages
are not supposed to ever pile up, this should be fine.
If messages pile up, then something is badly wrong somewhere, anyway.

Change-Id: Id8866998218a33f0d1b1fecc990b736a6f35f184
Signed-off-by: Ukri Niemimuukko <ukri.niemimuukko@intel.com>