[PORT FROM R44D] libcamera2: write Makernote data to EXIF APP2 segment
The Makernote data is currently written in the EXIF APP1 segment,
which is the standard EXIF location for such data. But the EXIF
APPn segment size restriction is roughly 64KB, which will
prevent us from using larger Makernote sizes in the future.

As a remedy, the Intel Makernote is now written to a separate
APP2 segment, and split across multiple APP2 segments, when the
size exeeds the limit of 64KB.

Change-Id: I95c2c1a38f6a84ed3634310ea9d6f0b95e5deeb5
Tracked-On: https://jira01.devtools.intel.com/browse/IMINAN-743
Signed-off-by: Lassi Väätämöinen <lassi.i.vaatamoinen@intel.com>
Reviewed-on: https://android.intel.com:443/256870