libcamera2: fix auto-focus sequence for CAF
The AF sequence handling for continuous AF-mode was incomplete.
This was hidden by the fact that no app in the past had actually
issued autoFocus() API call, when we were in continuous-picture
focus mode.

This patch addresses the issue in following cases, when autoFocus()
call is received from the framework/app:
- CAF *not* busy: return AF status immediately via callback
- CAF is busy: run AF until focused, send AF callback

Also, the AF is now locked after completing autoFocus() call in
continus AF mode. This means sending no Camera.AutoFocusMoveCallback
when the AF is locked.

Change-Id: I1ac99df3b7a283cb61afbe627b65c36d5716fa7c
Tracked-On: https://jira01.devtools.intel.com/browse/IMINAN-1822
Signed-off-by: Lassi Väätämöinen <lassi.i.vaatamoinen@intel.com>
Reviewed-on: https://android.intel.com:443/261760