libcamera2: force supported flash modes list to "off" only for Intel API
The GMS application used in L-dessert does not check for supported flash mode
changes. As we force the flash mode support list to contain only "off" in low-battery
condition, the following parameter validation will return error, as the application
does not change the flash mode setting accroding to the support list change.

Change-Id: I31a01e277489e90b087a2ef18d51846513d125d8
Tracked-On: https://jira01.devtools.intel.com/browse/IMINAN-2343
Signed-off-by: Lassi Väätämöinen <lassi.i.vaatamoinen@intel.com>
Reviewed-on: https://android.intel.com:443/290490