libcamera2: allocate HDR main buffer from graphics
With the devices having 4GB system memory the main buffer for
HDR has to be reserved under 2GB address space as the HW encoder
does not work otherwise. To comply with this restriction buffer is
reserved from the graphics.

Devices having 2GB or less system memory the allocation is made from
the heap instead of the framework memory.

Change-Id: I17efce020dfb9709f8e6ffa0eb5cbbc2a7671564
Tracked-On: https://jira01.devtools.intel.com/browse/IMINAN-8987
Signed-off-by: Tommi Franttila <tommi.franttila@intel.com>
Reviewed-on: https://android.intel.com:443/301622