[WORKAROUND]libcamera2: cap EV compensation
The GMS application used in L-dessert does not check for valid exposure
compensation values, nor does it handle errors when trying to set an
incorrect value.

Return success even when an incorrect value is passed in, but silently
ensure that the value is within the valid range.

Since the value is silently corrected in the validateParameter function
we also have to change it to no longer pass the param struct as const.

Change-Id: Ib3ac0a5060b826e9cf15917a54560593bc818ca2
Tracked-On: https://jira01.devtools.intel.com/browse/IMINAN-3526
Signed-off-by: David Weinehall <david.weinehall@intel.com>
Reviewed-on: https://android.intel.com:443/301620