[PORT FROM R44D] libcamera2: GPU warping initialization fix
GPU warper wasn't initialized when ULL was disabled but
trigger flag was set. Once when ULL was enabled GPU warper
won't be initialized because there were no change in the
trigger status, even if it should be initialized in case the
trigger was set.

This patch fixes the issue and sets the trigger flag properly.

Change-Id: I6c097d71477259f06aef7ac7b3b77199c1e0014e
Tracked-On: https://jira01.devtools.intel.com/browse/IMINAN-9571
Signed-off-by: Dalibor Segan <dalibor.segan@intel.com>
Signed-off-by: Lassi Väätämöinen <lassi.i.vaatamoinen@intel.com>
Reviewed-on: https://android.intel.com:443/305445