libcamera2: fix cts testPreviewPictureSizesCombination
When allocating snapshot buffer for image capture, it will skip
allocation if previous buffer could be reuse. However, it only
check width and height but not bpl, which will result in capture
fail when bpl and width are not the same. Fixed by taking snaphsot
buffer's bpl into consideration when deciding skipping allocation or
not. Also, remove unneccessary bpl calculation in allocateInputBuffer.

Change-Id: I43c7c4558935084d3313aea4de22d260ec48cfcc
Tracked-On: https://jira01.devtools.intel.com/browse/IMINAN-13114
Signed-off-by: Huang, Hayden <hayden.huang@intel.com>
Signed-off-by: Tommi Franttila <tommi.franttila@intel.com>
Reviewed-on: https://android.intel.com:443/311601