[WORKAROUND] libcamera2: call abort() in PreviewThread "livelock"
In case gralloc buffer lock/unlock operations or queue/dequeue from
preview window fail constantly, we can assume that the preview surface
is gone. This kind of situation may be introduced when lowmemorykiller
kills the application that holds the camera preview, for example.

If the camera application and preview surface(s) have died in a way
that camera framework has not been able to detect client
disconnection, and hence has failed to shutdown the HAL properly, as
a last resort camera HAL can do an abort() call. This way the camera
service will be in a consistent state with app and HAL eventually.
The downside is that all mediaserver services will be killed as well.

Since we have no known method of telling whether the preview window
is alive, we have to rely on a safe-period type of detection.
This patch uses fail count of 100, which in 30FPS preview case would
mean, e.g., 100 frames queuing errors(3 seconds of preview failure)
before abort() is called.

Change-Id: I388b2ac8a248d5e0c4bcfb24edfe4975f8986ce2
Tracked-On: https://jira01.devtools.intel.com/browse/IMINAN-13110
Signed-off-by: Lassi Väätämöinen <lassi.i.vaatamoinen@intel.com>
Signed-off-by: Tommi Franttila <tommi.franttila@intel.com>
Reviewed-on: https://android.intel.com:443/311602