libcamera2: provide correct statistics to 3A library
Sometimes the statistics read from driver wouldn't be
the lastest one. Provided out-of-date statistics to 3A,
which caused invalid AE handling.

This patch addresses the issue by dequeuing all the old
statistics out of the driver queue, based on exposure ID in
the statistic event. Statistics are read out of the driver queue,
until we get the statistics with matching exposure ID.
If HAL failed to dequeue the requested statistics,
the most close to requested statistics would be passed to 3A.

Depends on: Id492c169da3b6bdd3623a26364c08bb9dfa281a3

Change-Id: I144b746b234dbb8e00ec8f7b5a8eb4f32fa175c3
Tracked-On: https://jira01.devtools.intel.com/browse/IMINAN-13242
Signed-off-by: Xia ShouYong <shouyongx.xia@intel.com>
Signed-off-by: Tommi Franttila <tommi.franttila@intel.com>
Reviewed-on: https://android.intel.com:443/326323